requirements.txt
-e .
aws-cdk.aws_s3
aws-cdk.core
